+++ R/bootstrap.R
@@ -88,16 +88,20 @@ renv_bootstrap_download <- function(version) {
   nv <- numeric_version(version)
   components <- unclass(nv)[[1]]
 
-  methods <- if (length(components) == 4L) {
-    list(
+  # if this appears to be a development version of 'renv', we'll
+  # try to restore from github
+  dev <- length(components) == 4L
+
+  # begin collecting different methods for finding renv
+  methods <- c(
+    renv_bootstrap_download_tarball,
+    if (dev)
       renv_bootstrap_download_github
-    )
-  } else {
-    list(
+    else c(
       renv_bootstrap_download_cran_latest,
       renv_bootstrap_download_cran_archive
     )
-  }
+  )
 
   for (method in methods) {
     path <- tryCatch(method(version), error = identity)
@@ -231,6 +235,33 @@ renv_bootstrap_download_cran_archive <- function(version) {
 
   message("FAILED")
   return(FALSE)
+
+}
+
+renv_bootstrap_download_tarball <- function(version) {
+
+  # if the user has provided the path to a tarball via
+  # an environment variable, then use it
+  tarball <- Sys.getenv("RENV_BOOTSTRAP_TARBALL", unset = NA)
+  if (is.na(tarball))
+    return()
+
+  # allow directories
+  info <- file.info(tarball, extra_cols = FALSE)
+  if (identical(info$isdir, TRUE)) {
+    name <- sprintf("renv_%s.tar.gz", version)
+    tarball <- file.path(tarball, name)
+  }
+
+  # bail if it doesn't exist
+  if (!file.exists(tarball))
+    return()
+
+  fmt <- "* Bootstrapping with tarball at path '%s'."
+  msg <- sprintf(fmt, tarball)
+  message(msg)
+
+  tarball
 
 }
 
@@ -692,3 +723,4 @@ renv_bootstrap_user_dir_impl <- function(path) {
     "~/.cache/R/renv"
 
 }
+