+++ R/r.R
@@ -159,16 +159,19 @@ r_cmd_install <- function(package, path, ...) {
   if (renv_platform_macos() && renv_package_type(path) == "source")
     renv_xcode_check()
 
-  # perform platform-specific preinstall checks
+  # perform platform-specific pre-install checks
   renv_scope_install()
 
   # perform the install
+  # note that we need to supply '-l' below as otherwise the library paths
+  # could be changed by, for example, site-specific profiles
   args <- c(
     "--vanilla",
     "CMD", "INSTALL", "--preclean", "--no-multiarch",
     r_cmd_install_option(package, "configure.args", TRUE),
     r_cmd_install_option(package, "configure.vars", TRUE),
     r_cmd_install_option(package, c("install.opts", "INSTALL_opts"), FALSE),
+    "-l", shQuote(library),
     ...,
     shQuote(path)
   )